+++ src/crontab_lib/errors/invalid_cron_format_error.py
@@ -1,15 +1,14 @@
 """This exception represents an error if a cron job has a bad syntax."""
-from crontab_lib import CronJob
 
 
 class InvalidCronFormatError(Exception):
     """This exception represents an error if a cron job has a bad syntax."""
 
-    def __init__(self, cron_job: CronJob):
+    def __init__(self, cron_job: str):
         """Initialize the InvalidCronFormatError.
 
         Args:
             cron_job: The problematic cron job.
         """
-        self.message = f"The con job ${str(cron_job)} is not valid"
+        self.message = f"The cron job ${cron_job} is not valid"
         super().__init__(self.message)