+++ src/crontab_lib/managers/inserters/__init__.py
@@ -1,1 +1,2 @@
 """Exposed inserters classes and methods."""
+from crontab_lib.managers.inserters.crontab_inserter import CrontabInserter